[REFACTOR] errors.rs file in error.rs
