+++ daemon/src/error.rs
