+++ daemon/src/services/cargo.rs
@@ -6,7 +6,7 @@ use nanocl_models::cargo::CargoPartial;
 use nanocl_models::generic::GenericNspQuery;
 
 use crate::utils;
-use crate::errors::HttpResponseError;
+use crate::error::HttpResponseError;
 use crate::models::Pool;
 
 /// Endpoint to create a new cargo
